avoid lombok for now
