+++ java8/pom.xml
@@ -20,14 +20,7 @@
             <artifactId>jackson-module-kotlin</artifactId>
             <version>2.5.1</version>
         </dependency>
-            <dependency>
-                <groupId>org.projectlombok</groupId>
-                <artifactId>lombok</artifactId>
-                <version>1.16.2</version>
-                <scope>provided</scope>
-            </dependency>
-
-        </dependencies>
+    </dependencies>
 
     <build>
         <plugins>
